Empty getDataSummary methods
