+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/datafeed/extractor/aggregation/CompositeAggregationDataExtractor.java
@@ -76,6 +76,11 @@ class CompositeAggregationDataExtractor implements DataExtractor {
     }
 
     @Override
+    public DataSummary getSummary() {
+        return null;
+    }
+
+    @Override
     public boolean hasNext() {
         return hasNext;
     }
