+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/datafeed/extractor/scroll/ScrollDataExtractor.java
@@ -68,6 +68,11 @@ class ScrollDataExtractor implements DataExtractor {
     }
 
     @Override
+    public DataSummary getSummary() {
+        return null;
+    }
+
+    @Override
     public boolean hasNext() {
         return hasNext;
     }
